backend/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
bleach==5.0.1
Django==4.0.6
django-cors-headers==3.13.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
webencodings==0.5.1
